requirements.txt
bcrypt==4.0.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
mysql-connector-python==9.0.0
numpy==1.24.1
pandas==2.2.2
passlib==1.7.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.31
uvicorn==0.30.1
aiomysql
asyncmy